+++ src/RcppExports.cpp
@@ -19,9 +19,51 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// build_examples2
+List build_examples2(NumericVector timeS, NumericVector lags, int nt);
+RcppExport SEXP _tsfgrnn_build_examples2(SEXP timeSSEXP, SEXP lagsSEXP, SEXP ntSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector >::type timeS(timeSSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type lags(lagsSEXP);
+    Rcpp::traits::input_parameter< int >::type nt(ntSEXP);
+    rcpp_result_gen = Rcpp::wrap(build_examples2(timeS, lags, nt));
+    return rcpp_result_gen;
+END_RCPP
+}
+// build_examples_m
+List build_examples_m(NumericVector timeS, NumericVector lags, int nt);
+RcppExport SEXP _tsfgrnn_build_examples_m(SEXP timeSSEXP, SEXP lagsSEXP, SEXP ntSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector >::type timeS(timeSSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type lags(lagsSEXP);
+    Rcpp::traits::input_parameter< int >::type nt(ntSEXP);
+    rcpp_result_gen = Rcpp::wrap(build_examples_m(timeS, lags, nt));
+    return rcpp_result_gen;
+END_RCPP
+}
+// build_examples_a
+List build_examples_a(NumericVector timeS, NumericVector lags, int nt);
+RcppExport SEXP _tsfgrnn_build_examples_a(SEXP timeSSEXP, SEXP lagsSEXP, SEXP ntSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector >::type timeS(timeSSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type lags(lagsSEXP);
+    Rcpp::traits::input_parameter< int >::type nt(ntSEXP);
+    rcpp_result_gen = Rcpp::wrap(build_examples_a(timeS, lags, nt));
+    return rcpp_result_gen;
+END_RCPP
+}
 
 static const R_CallMethodDef CallEntries[] = {
     {"_tsfgrnn_regression_2", (DL_FUNC) &_tsfgrnn_regression_2, 4},
+    {"_tsfgrnn_build_examples2", (DL_FUNC) &_tsfgrnn_build_examples2, 3},
+    {"_tsfgrnn_build_examples_m", (DL_FUNC) &_tsfgrnn_build_examples_m, 3},
+    {"_tsfgrnn_build_examples_a", (DL_FUNC) &_tsfgrnn_build_examples_a, 3},
     {NULL, NULL, 0}
 };