+++ tests/testthat/test_grnn_model.R
@@ -1,6 +1,7 @@
-context("Building a GRNN model")
 
-expect_error(grnn_model(ts(1:5), lags = 0:3, sigma = 5))
+expect_error(grnn_model(ts(1:5), lags = 0:3,
+                        sigma = 5, transform = "none"))
 
-expect_error(grnn_model(ts(1:5), lags = 3:5, sigma = 1),
+expect_error(grnn_model(ts(1:5), lags = 3:5,
+                        sigma = 1, transform = "none"),
              "Impossible to create one example")